refactor: change naming convention for pinia stores & composables (#135)
