+++ src/views/GameView.vue
@@ -2,7 +2,7 @@
 import GameCanvas from '@/components/ui/layout/game/GameCanvas.vue'
 import SideBar from '@/components/ui/layout/game/sidebar/SideBar.vue'
 import TopBar from '@/components/ui/layout/game/topbar/TopBar.vue'
-import useSelectedBuildingArea from '@/composables/useSelectedBuildingArea.js'
+import { useSelectedBuildingAreaStore } from '@/composables/useSelectedBuildingAreaStore.js'
 import { ref } from 'vue'
 
 const sidebarOpen = ref(false)
@@ -10,7 +10,7 @@ function toggleSidebar() {
   sidebarOpen.value = !sidebarOpen.value
 }
 
-const selectedBuildingArea = useSelectedBuildingArea()
+const selectedBuildingArea = useSelectedBuildingAreaStore()
 selectedBuildingArea.$subscribe((_, newId) => {
   if (newId) {
     // The user selected a building area.
